Make input_url possible as positional argument, cleanup factory tests (#1321)
* Make input_url possible as positional argument, cleanup factory tests

* Now throws TraitError
