+++ ctapipe/io/eventsource.py
@@ -7,7 +7,7 @@ from traitlets.config.loader import LazyConfigValue
 from ctapipe.core import Component, non_abstract_children, ToolConfigurationError
 from ctapipe.core import Provenance
 from ctapipe.core.plugins import detect_and_import_io_plugins
-from ctapipe.core.traits import Path, Int, Set, TraitError
+from ctapipe.core.traits import Path, Int, Set
 
 __all__ = ["EventSource", "event_source"]
 
@@ -121,7 +121,7 @@ class EventSource(Component):
         )
     ).tag(config=True)
 
-    def __init__(self, config=None, parent=None, **kwargs):
+    def __init__(self, input_url=None, config=None, parent=None, **kwargs):
         """
         Class to handle generic input files. Enables obtaining the "source"
         generator, regardless of the type of file (either hessio or camera
@@ -139,7 +139,7 @@ class EventSource(Component):
             Set to None if no Tool to pass.
         kwargs
         """
-        super().__init__(config=config, parent=parent, **kwargs)
+        super().__init__(config=config, parent=parent, input_url=input_url, **kwargs)
 
         self.metadata = dict(is_simulation=False)
         self.log.info(f"INPUT PATH = {self.input_url}")
@@ -283,6 +283,10 @@ class EventSource(Component):
         if input_url == "" or input_url is None:
             raise ToolConfigurationError("EventSource: No input_url was specified")
 
+        # validate input url with the traitel validate method
+        # to make sure it's compatible and to raise the correct error
+        input_url = EventSource.input_url.validate(obj=None, value=input_url)
+
         detect_and_import_io_plugins()
         available_classes = non_abstract_children(cls)
 
@@ -329,6 +333,5 @@ class EventSource(Component):
 
         if isinstance(config.EventSource.input_url, LazyConfigValue):
             config.EventSource.input_url = cls.input_url.default_value
-        elif not isinstance(config.EventSource.input_url, str):
-            raise TraitError("Wrong type specified for input_url traitlet")
+
         return event_source(config.EventSource.input_url, config=config, **kwargs)
