+++ ctapipe/io/simteleventsource.py
@@ -151,7 +151,7 @@ class SimTelEventSource(EventSource):
         ),
     ).tag(config=True)
 
-    def __init__(self, config=None, parent=None, gain_selector=None, **kwargs):
+    def __init__(self, input_url, config=None, parent=None, gain_selector=None, **kwargs):
         """
         EventSource for simtelarray files using the pyeventio library.
 
@@ -169,7 +169,7 @@ class SimTelEventSource(EventSource):
             The GainSelector to use. If None, then ThresholdGainSelector will be used.
         kwargs
         """
-        super().__init__(config=config, parent=parent, **kwargs)
+        super().__init__(input_url=input_url, config=config, parent=parent, **kwargs)
         self._camera_cache = {}
 
         self.file_ = SimTelFile(
